login and register

+++ src/app/components/calendar/calendar.component.ts
@@ -25,6 +25,7 @@ import flatpickr from 'flatpickr';
 import { PlannerService } from 'src/app/services/planner.service';
 import { EventObj } from '../../models/EventObj';
 import { ToastrService } from 'ngx-toastr';
+import { LocalStorageService } from 'src/app/services/local-storage.service';
 
 flatpickr.l10ns.default.firstDayOfWeek = 1;
 
@@ -98,7 +99,8 @@ export class CalendarComponent implements OnInit {
   constructor(
     private modal: NgbModal,
     private plannerService: PlannerService,
-    private toastr: ToastrService
+    private toastr: ToastrService,
+    private localStorage: LocalStorageService
   ) {}
 
   ngOnInit(): void {
@@ -178,7 +180,7 @@ export class CalendarComponent implements OnInit {
     this.eventObjToSend = {
       id: 1,
       calendarEvent: this.tempEvent,
-      userID: 1,
+      userID: this.localStorage.getUserIdFromLocalStorage(),
       link: '',
     };
 
@@ -222,13 +224,15 @@ export class CalendarComponent implements OnInit {
   }
 
   deleteEvent(eventToDelete: CalendarEvent): void {
-    this.events = this.events.filter(
-      (event) => event.calendarEvent !== eventToDelete
-    );
-    this.plannerService.deleteEvent(eventToDelete).subscribe(() => {
-      this.updateEventsList();
-      this.toastr.success(eventToDelete.title, 'Deleted');
-    });
+    if (eventToDelete.id != null) {
+      this.events = this.events.filter(
+        (event) => event.calendarEvent !== eventToDelete
+      );
+      this.plannerService.deleteEvent(eventToDelete).subscribe(() => {
+        this.updateEventsList();
+        this.toastr.success('Event deleted', 'Success');
+      });
+    }
 
     this.tempEvent = null;
     this.clickedEdit = false;
